+++ src/estudiante/estudiante.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateEstudianteDto } from './dto/create-estudiante.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Estudiante } from "./entities/estudiante.entity";
 import { UpdateEstudianteDto } from './dto/update-estudiante.dto';
+
 
 
 @Injectable()
@@ -13,7 +14,16 @@ export class EstudianteService {
   constructor(@InjectRepository(Estudiante) private readonly estudianteRepository: Repository<Estudiante>) { }
 
 
-  createEstudiante(estudiante: CreateEstudianteDto) {
+  async createEstudiante(estudiante: CreateEstudianteDto) {
+    const ciudadFound = await this.estudianteRepository.findOne({
+      where: {
+        nombre: estudiante.nombre
+      }
+    })
+
+    if (ciudadFound) {
+      return new HttpException('La ciudad ya existe', HttpStatus.CONFLICT)
+    }
     const newEstudiante = this.estudianteRepository.create(estudiante)
     return this.estudianteRepository.save(newEstudiante);
   }
@@ -23,17 +33,34 @@ export class EstudianteService {
 
   }
 
-  findOne(id: number) {
-    return this.estudianteRepository.findOne({ where: { id } });
+  async findOne(id: number) {
+    const ciudadFound = await this.estudianteRepository.findOne({ where: { id } });
+
+    if (!ciudadFound) {
+      return new HttpException('Ciudad no encontrada', HttpStatus.NOT_FOUND);
+    }
+    return ciudadFound;
   }
+  
 
-  update(id: number, updateEstudiante: UpdateEstudianteDto) {
-    return this.estudianteRepository.update({ id }, updateEstudiante)
 
+  async update(id: number, update: UpdateEstudianteDto) {
+    const estudianteFound = await this.estudianteRepository.findOne({ where: { id } });
+
+    if(!estudianteFound){
+      return new HttpException('Estudiante no encontrado', HttpStatus.NOT_FOUND);
+    }
+    const updateEstudiante = Object.assign(estudianteFound, update);
+    return this.estudianteRepository.save(updateEstudiante)
   };
 
 
-  remove(id: number) {
-    return this.estudianteRepository.delete({ id });
+  async remove(id: number) {
+    const estudianteFound = await this.estudianteRepository.findOne({ where: { id } });
+ 
+    if(!estudianteFound){
+      return new HttpException('Estudiante no encontrado', HttpStatus.NOT_FOUND);
+    }
+    return this.estudianteRepository.delete({id});
   }
 }
